Fix Jenkins pipeline
Jenkins xunit plugin was complaining about a missing Dart configuration
file. This file is only generated when fully preparing CTest to work
with CDash, which happens only when using `include(CTest)` instead of
`enable_testing()`.

The cmakebuild plugin was configuring using CMake twice, once for the
configuration and once for the build stage. The build stage reset the
`ENABLE_TESTING` flag enabled by the build stage. The plugin now uses
the `--build` argument to simply build and not configure the project.